update BetterLED to use a costom command Scheduler
So many changes have been made, it now uses a fully diffrent system taking in LED Commands that are special classes similar to wpilib's classes but simpler, so that the implementation is simpler.

The only problem is interrupting a task is impossible so there is an option to force remove elements from the call list, as the scheduler WILL wait for that command to finish.