add a mutex for led thread system
Concurrency is now correctly impemented, this should improve the systems.

For the Boolean that kills the thread, I just made it volatile because it will not be changed by the thread in any situation, it will only be read by it.

Also even if it reads it at the same time, its fine cause it can read it in the next iteration of the infinite for loop.